paperenvbench/tasks/mae_reconstruction_minimal/requirements_lock.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3
Pygments==2.20.0
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
filelock==3.29.0
fsspec==2026.4.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.14.0
idna==3.15
markdown-it-py==4.2.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==26.2
pillow==12.2.0
rich==15.0.0
setuptools==82.0.1
shellingham==1.5.4
sympy==1.14.0
timm==0.6.13
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.3
triton==3.0.0
typer==0.25.1
typing_extensions==4.15.0
wheel==0.47.0
